requirements.txt
numpy==1.16.6
opencv-python==4.4.0.46
pillow==8.0.1
pycocotool==2.0.2
torch==1.8.1
torchvision==0.9.1
terminaltables==3.1.0
wget==3.2












